requirements.txt
streamlit==1.32.0
python-dotenv==1.0.1
groq==0.4.2
pandas==2.2.1
numpy==1.26.4
plotly==5.19.0
python-docx==1.1.0 